Source base AMI from configuration file
Instead of having the base AMIs that are used to build the Stardog
AMIs baked into the code we will get them from a configuration file
~/.graviton/base-amis-<sd version>.json.  If that file is not found
it is created from the values originally baked into the code.  This
gives users a chance to override and update those values.  This patch
also updates the default base AMI to Ubuntu 16.04 hvm.