Adding in version control fro dependent executables
graviton currently depends on two single file executables, terraform
and packer.  Both of these programs are single executables (waving
hands at some new weirdness in terraform) so they can be copied and
placed into ~/.graviton and thereby we can better control the versions
used on a given system.  Recently terraform had a breaking change
and thus users that upgraded terraform broke their graviton installation.

This patch locks down versions of terraform and packer to specfic releases.
It first looks for the executables under ~/.gravtion.  If they are there
and have the correct `--version` output then they are used.  Otherwise the
system path is searched.  If found it is copied to ~/.graviton.  Otherwise
They are downloaded from a known location.