Adding support for a custom script
This adds an option that allows a user to supply a custom script to
run on all Stardog nodes.  Just the script itself is copied to the
remote machines thus any dependecies that it has must also be on that
machine.  This patch also adds /mnt/data/stardog-home/logs to the
recursive list of directories that are pulled down in the tarball
with the logs command.